refactor: move all interfaces and types into the associated module
style types still solo since they are used by many modules
