+++ terminus-ui/src/card/card.component.ts
@@ -7,9 +7,17 @@ import {
   ElementRef,
 } from '@angular/core';
 
-import {
-  TsAspectRatioTypes,
-} from './../utilities/types';
+
+/**
+ * Define the allowed aspect ratios. Used in {@link TsCardComponent}
+ */
+export type TsAspectRatioTypes =
+  '16:9'
+  | '4:3'
+  | '3:2'
+  | '5:4'
+  | '1:1'
+;
 
 
 /**
