+++ terminus-ui/src/date-range/date-range.component.ts
@@ -13,8 +13,23 @@ import {
 import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { MatDatepickerInputEvent } from '@angular/material/datepicker';
 
-import { TsDateRange } from './../utilities/interfaces/date-range.interface';
 import { TsStyleThemeTypes } from './../utilities/types/style-theme.types';
+
+
+/**
+ * Define the structure of the date range object used by {@link TsDateRangeComponent}
+ */
+export interface TsDateRange {
+  /**
+   * The start date of the range
+   */
+  start: Date | null;
+
+  /**
+   * The end date of the range
+   */
+  end: Date | null;
+}
 
 
 /**
