+++ terminus-ui/src/menu/menu.component.ts
@@ -11,12 +11,34 @@ import {
 } from '@angular/core';
 import { MatMenuTrigger } from '@angular/material/menu';
 
-import {
-  TsMenuPositionTypesX,
-  TsMenuPositionTypesY,
-  TsStyleThemeTypes,
-  TsMenuTriggerTypes,
-} from './../utilities/types';
+import { TsStyleThemeTypes } from './../utilities/types/style-theme.types';
+
+
+/**
+ * Define the allowed X positions for a menu
+ */
+export type TsMenuPositionTypesX =
+  'before'
+  | 'after'
+;
+
+
+/**
+ * Define the allowed Y positions for a menu
+ */
+export type TsMenuPositionTypesY =
+  'above'
+  | 'below'
+;
+
+
+/**
+ * Define the allowed trigger types for a menu
+ */
+export type TsMenuTriggerTypes
+  = 'default'
+  | 'utility'
+;
 
 
 /**
