+++ terminus-ui/src/spacing/vertical-spacing.directive.ts
@@ -6,7 +6,25 @@ import {
 } from '@angular/core';
 
 import { TS_SPACING } from './spacing.constant';
-import { TsVerticalSpacingTypes } from './../utilities/types';
+
+
+/**
+ * Define the accepted string values for {@link TsVerticalSpacingDirective}
+ */
+export type TsVerticalSpacingTypes =
+  'small--2' |
+  'small--1' |
+  'small--0' |
+  'none' |
+  'default--0' |
+  'large--0' |
+  'large--1' |
+  'large--2' |
+  'large--3' |
+  'large--4' |
+  'large--5' |
+  'large--6'
+;
 
 
 /**
