+++ terminus-ui/src/tooltip/tooltip.component.ts
@@ -5,7 +5,18 @@ import {
   ViewEncapsulation,
 } from '@angular/core';
 
-import { TsTooltipPositionTypes } from './../utilities/types';
+
+/**
+ * Define the accepted string values for the {@link TsTooltipComponent} position
+ */
+export type TsTooltipPositionTypes =
+  'left' |
+  'right' |
+  'above' |
+  'below' |
+  'before' |
+  'after'
+;
 
 
 /**
