--- terminus-ui/src/utilities/interfaces/date-range.interface.ts
@@ -1,14 +1,0 @@
-/**
- * Define the structure of the date range object used by {@link TsDateRangeComponent}
- */
-export interface TsDateRange {
-  /**
-   * The start date of the range
-   */
-  start: Date | null;
-
-  /**
-   * The end date of the range
-   */
-  end: Date | null;
-}
