--- terminus-ui/src/utilities/interfaces/index.ts
@@ -1,6 +1,0 @@
-export * from './login-form-response.interface';
-export * from './paginator.interface';
-export * from './navigation-item.interface';
-export * from './navigation-payload.interface';
-export * from './user.interface';
-export * from './search.interface';
