--- terminus-ui/src/utilities/interfaces/navigation-payload.interface.ts
@@ -1,16 +1,0 @@
-/**
- * Define the expected response from the {@link TsNavigationComponent} emitter
- */
-export interface TsNavigationPayload {
-  /**
-   * The mouse click event
-   */
-  event: MouseEvent;
-
-  /**
-   * The selected item
-   */
-  action: {
-    type: string;
-  };
-}
