--- terminus-ui/src/utilities/interfaces/search.interface.ts
@@ -1,9 +1,0 @@
-/**
- * Define the user object interface
- */
-export interface TsSearchResponse {
-  /**
-   * The search query
-   */
-  query: string;
-}
