--- terminus-ui/src/utilities/interfaces/user.interface.ts
@@ -1,9 +1,0 @@
-/**
- * Define the user object interface
- */
-export interface TsUser {
-  /**
-   * The user's full name
-   */
-  fullName: string;
-}
