--- terminus-ui/src/utilities/interfaces/validator-response.interface.ts
@@ -1,14 +1,0 @@
-/**
- * Define the structure of the validation response object
- */
-export interface ValidationResponse {
-  /**
-   * The expected object structure
-   */
-  [key: string]: {
-    /**
-     * The expected validation result
-     */
-    valid: boolean;
-  };
-}
