--- terminus-ui/src/utilities/regex/credit-card.regex.ts
@@ -1,4 +1,0 @@
-/**
- * Define the error message for an invalid credit card
- */
-export const creditCardMessage = `Invalid credit card number.`;
