--- terminus-ui/src/utilities/regex/email.regex.ts
@@ -1,4 +1,0 @@
-/**
- * Define the error message for an invalid email
- */
-export const emailMessage = `Invalid email address.`;
