--- terminus-ui/src/utilities/regex/password.regex.ts
@@ -1,4 +1,0 @@
-/**
- * Define the error message for an invalid password
- */
-export const passwordMessage = `Password must be between 6 and 100 characters, and contain a number.`;
