--- terminus-ui/src/utilities/types/aspect-ratio.types.ts
@@ -1,10 +1,0 @@
-/**
- * Define the allowed aspect ratios. Used in {@link TsCardComponent}
- */
-export type TsAspectRatioTypes =
-  '16:9'
-  | '4:3'
-  | '3:2'
-  | '5:4'
-  | '1:1'
-;
