--- terminus-ui/src/utilities/types/autocomplete.types.ts
@@ -1,10 +1,0 @@
-/**
- * Define a type for allowed {@link TsAutocompleteComponent} formatter function
- */
-export type TsAutocompleteFormatterFn = (value: any) => string;
-
-
-/**
- * Define a type for allowed {@link TsAutocompleteComponent} comparator function
- */
-export type TsAutocompleteComparatorFn = (value: any) => string;
