--- terminus-ui/src/utilities/types/index.ts
@@ -1,6 +1,0 @@
-export * from './aspect-ratio.types';
-export * from './autocomplete.types';
-export * from './menu.types';
-export * from './spacing.types';
-export * from './style-theme.types';
-export * from './tooltip.types';
