--- terminus-ui/src/utilities/types/input.types.ts
@@ -1,1 +1,0 @@
-
