--- terminus-ui/src/utilities/types/menu.types.ts
@@ -1,25 +1,0 @@
-/**
- * Define the allowed X positions for a menu
- */
-export type TsMenuPositionTypesX =
-  'before'
-  | 'after'
-;
-
-
-/**
- * Define the allowed Y positions for a menu
- */
-export type TsMenuPositionTypesY =
-  'above'
-  | 'below'
-;
-
-
-/**
- * Define the allowed trigger types for a menu
- */
-export type TsMenuTriggerTypes
-  = 'default'
-  | 'utility'
-;
