--- terminus-ui/src/utilities/types/sort.types.ts
@@ -1,8 +1,0 @@
-/**
- * Define the allowed sort directions for {@link TsSort}
- */
-export type TsSortDirection
-  = 'asc'
-  | 'desc'
-  | ''
-;
