--- terminus-ui/src/utilities/types/spacing.types.ts
@@ -1,17 +1,0 @@
-/**
- * Define the accepted string values for {@link TsVerticalSpacingDirective}
- */
-export type TsVerticalSpacingTypes =
-  'small--2' |
-  'small--1' |
-  'small--0' |
-  'none' |
-  'default--0' |
-  'large--0' |
-  'large--1' |
-  'large--2' |
-  'large--3' |
-  'large--4' |
-  'large--5' |
-  'large--6'
-;
