--- terminus-ui/src/utilities/types/tooltip.types.ts
@@ -1,11 +1,0 @@
-/**
- * Define the accepted string values for the {@link TsTooltipComponent} position
- */
-export type TsTooltipPositionTypes =
-  'left' |
-  'right' |
-  'above' |
-  'below' |
-  'before' |
-  'after'
-;
